+++ src/JsonParser.php
@@ -19,10 +19,17 @@ use yii\helpers\Json;
 class JsonParser extends BaseObject implements ParserInterface
 {
     /**
+     * @var bool whether to return objects in terms of associative arrays.
+     * @since 2.0.8
+     */
+    public $asArray = true;
+
+
+    /**
      * {@inheritdoc}
      */
     public function parse(Response $response)
     {
-        return Json::decode($response->getContent());
+        return Json::decode($response->getContent(), $this->asArray);
     }
 }